-refactored the entire code that is responsible for fetching and inserting comments -refactored Comment DTO and Comment classes -Added CommentMapper for converting DTO Comment objects to Comment objects

+++ app/src/main/java/com/example/bookapp/viewModels/ViewModelPost.kt
@@ -30,7 +30,7 @@ class ViewModelPost(application: Application) : AndroidViewModel(application) {
         return postRepository.myPosts
     }
 
-    fun getPostByID(id: Long): LiveData<Post> {
+    fun getPostByID(id: Int): LiveData<Post> {
         viewModelScope.launch {
             postRepository.fetchPostByID(id)
         }
